Dashboards section (incomplete)
I just need to organize the dashboards in a better way that fit in the screen, and also I need to fixed the JS script. I created a JS code to export the dashboard data to excel sheet, but when you open the downloaded file, there is nothing. I need to fix this.